LPD-84722 Adapt component to response type
