+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/js/contacts/pages/account/utils/types.ts
@@ -1,0 +1,11 @@
+import {TrendClassification} from 'segment/types';
+
+export type Metric = {
+	trend: Trend;
+	value: number;
+};
+
+export type Trend = {
+	percentage: number;
+	trendClassification: TrendClassification;
+};
